fix: RootLayout 관련 파일 정리 및 단일 레이아웃 파일로 통합
브레이크포인트별 레이아웃 분리로 인한 탭 상태 초기화 문제를 해결하기 위해, 레이아웃 파일을 하나로 통합하고 기존 파일들을 삭제했습니다.
