+++ frontend/src/components/common/Navbar/NavbarWrapper.tsx
@@ -1,17 +1,19 @@
-// import { useMediaQuery } from 'react-responsive'
-// import { NavbarDesktop } from './NavbarDesktop'
-// import { NavbarTablet } from './NavbarTablet'
-// import { NavbarMobile } from './NavbarMobile'
+import { NavbarDesktop } from './NavbarDesktop'
+import { NavbarMobile } from './NavbarMobile'
+import { NavbarTablet } from './NavbarTablet'
 
-// export const NavbarWrapper = () => {
-//     const isDesktop = useMediaQuery({ minWidth: 1440 })
-//     const isTablet = useMediaQuery({ minWidth: 768, maxWidth: 1439 })
-
-//     if (isDesktop) {
-//         return <NavbarDesktop />
-//     } else if (isTablet) {
-//         return <NavbarDesktop />
-//     } else {
-//         return <NavbarDesktop />
-//     }
-// }
+export const NavbarWrapper = () => {
+    return (
+        <>
+            <div className="hidden desktop:block">
+                <NavbarDesktop />
+            </div>
+            <div className="hidden tablet:block desktop:hidden">
+                <NavbarTablet />
+            </div>
+            <div className="block tablet:hidden">
+                <NavbarMobile />
+            </div>
+        </>
+    )
+}
